Initial Resources for State Machine and State Machine Demo
The state machine and general software architecture is outlined by this commit.
The state machine implementation uses states which are implementations of a
common interface. Alongside the state machine is a bank containing all data for
and object type. The states have tranisition information as well as actions that
act on data. Action behavior is defined by unique classes called cartridges.
Cartridges take data in, edit it, then return it to the state. The state will
then pass the data back to the data bank, where it is then used by the engine
for display and calculation.

This implementation ensures all numbers are crunched before anything is
applied to the engine, keeping everything from being tangled in engine
calls.