State Machine updated to express state changes at the most basic level
The State Machine can now handle transitioning states, which can be verified by
pressing the space bar while the game is running. The result transitions
between two states. However, the code currently calls too many constructors so
a refactor will be coming that places the States as single instances inside the
state machine.

Another issue to be solved going forward is how to integrate deltaTime into the
acceleration function. Either in the cartridge or in the actual translation call