Added rudimentary updating of engine based on state-cartridge-data system
The state-cartridge-data system now interfaces with the engine in the PlayerController. It expects all of the data to be updated before the engine is notified, then any applicable actions taken at the Unity level are performed. Currently experiencing a bug in which the object moves backward constantly, but this is easily mitigated by differentiating between "no input" state" and "no input with no velocity" state. The cartridges still need to be managed by the state machine, so that will be coming next.