States and Cartridges are now all delivered by the master object controller.
Previously, States and Cartridges were newed whenever a state changed to another state. Now, these objects are now passed from the Controller to the State
Machine. This should ensure that there is minimal memory management going on
once an entire scene is loaded. It appears to function exactly the same. Also
included is the first of many debug functions: quitting the game!