Updated states to ensure that direction of travel is preserved.
The cube now controls much more like expected. A persistent issue is the cart
for handling returns a positive angle while also not taking the magnitude of the
input axis into account. The next step is to figure out an elegant solution for exposing input to the state machine while preserving the architecture of the
system.