Implemented solution to pulling data in from the engine and turning fixed.
A new function EnginePull() has been added to the PlayerController. As this
class is the intended interface with the rest of the Unity Engine, it was the
right move to have a function pull data in from the engine for use by the states
and then pushed back out as it has been done so far. This implementation has
the data structure contain engine information such as input, which can then
be served up to the cartridges to influence behavior. At the same time, this
implementation fixed the turning issue as that was taking in a single float, a
float now influenced by the input axis!