Camera can now track objects while remaining stationary. New revelations!
The camera now tracks a target cleanly. While the architecture still seems
relatively sound, it revealed that the data passed in through the state machine
essentially requires a unique state machine for each object. If there is a type
casting solution it will fix the issue, but currently each object has to have
its own version of StateMachine, Data, and Controller. However, considering the
data, controllers, and cartridges all need individual implementations, the
overhead is minimal and the debugging/feature expansion benefit is worth it.