Objects can now be oriented toward a surface on landing with the object
The simple test case of a box falling on a plane in any orientation now works.
The mising pieces are how to handle the "floating" that occurs when the point
of contact is not the center of the object. The likely solution is applying the
rotation AROUND the point of contact.