Heavily improved angle calculation and ground-air transitions
Several enhancements on the workflow for calculating surface direction have improved how the player interacts with the surfaces. The offset has been removed as quickly as it came in, thanks to a recommendation on how to handle checking underfoot. The downward raycast now has different set values for ground and air, improving transitions. The behavior on the ground has been refactored, removing the need for re-transitioning under the current test environment.