Fixed the unusual turnaround bug and started implementing crashes
The turnaround bug was caused by the way the character's engine representation takes data in. The data is now represented as a rotation and position in world space rather than using a rotation delta, making the rotations more consistent. There are still quirks to work out with downward angles and cliffs, but this works well. Crashes are the next priority, and this commit includes the first details on crashing.