Heavy Refactoring and Slowing Down
The game has gone too long without the ability to slow down and stop! This commit includes that change, though it is currently a work in progress as turning is not implemented with slowing down. This may lead to breaking turning out into a separate state machine. 

Also included in this commit is major refactoring, placing references to the data in individual states rather than getting passed to the controller. This should reduce reference passing and ultimately improve performance.