Refactoring! The state machines have been broken up, fixing prior bugs
The individual state machine has been refactored into a set of state machines that each govern a different part of the player object's behavior. This fixed the issue in the last commit of having to come up with a hierarchy for state machine commands to ensure slowing down and turning weren't order-dependent.