Jumping implementation has begun
This commit encompasses the beginning of development on a basic jump mechanic. The rough core is there but there are issues as there is no differentiation between a jump and a fall, so to speak. This creates issues when the jump is too weak to actually lift the player, causing floor clipping.