Started implementation of crashing
Crashing into a wall leads to the player stopping now, but crashing twice is iffy due to the hierarchy of state machine commands. The problem is either crash does not allow the player to return to a stationary state OR the timer remains above the limit and once recovered from a crash, the player will not crash again.