The halfpipe illuminated much needed updates to air to ground
The halfpipe showed the need for two-point rotation handling. In addition, when landing the velocity appears to be reset to a certain degree. The transitions between landing and hitting the air are not smooth. Ground2Air puts too much weight on the Y value, while Air2Ground doesn't want to update the velocity realistically. All in all there is a lot of work ahead to fix these.