Improved jumping, no longer a heavy disparity between up and down hill
Previously the scaling of jumping based on current angle was incorrect, proving moon jumps going uphill and tiny hops going downhill. The function was intended to be closer to reality, but it is not quite correct. However, it seems that jumps are uniform based on direction, which is an improvement.