Tricking State Machine has been added
To break up the monotony of working on the same parts of the camera and player controller, the beginnings of the trick system are encompassed by this commit. This includes a working state machine that otherwise does nothing. This will eventually necessitate animation to govern tricks being ready or not. This breaks the branching rules but I am the only developer so I make the git rules :)