Basic Trick System Implemented
The conversions between floats and integers is a code smell, but it's a smell that allows the tricking system to be time independent. I believe that is worth it. The Trick Dictionary concept did not work out, but I do believe there is useful work to be done in making the trick system easier to work with.