Basic Scoring System Implemented
Using the messaging system and the player controller, the scoring system can now go from player input all the way to incrementing the onscreen display. This has illuminated the need for enhancements to the messaging system and the way data is broken up within a single game entity. Either way, the basics have come together nicely.