Menu Improvements!
Menu actions have now been better articulated. Menu items can now be scaled and perform specific actions, such as loading scenes based on index. The static MenuSelectionData allows the currently loaded scene to choose the next level as well as data that will be pulled in on scene load (in the future!)