Zones added
This commit adds the functionality for a static zone lookup through the GameMasterController, which controls world state and universal elements that may be accessed by a wide variety of classes. Currently the data structures are updated, but meaningful messages on zone change are not sent. That will be added later as elements that may depend on zones get added.