Basic Timer implemented
This commit contains a very basic version of a timer. It simply ticks up. Adding new messages for pausing the game and reaching checkpoints (using the zone system we implemented earlier!) will allow for more intricate time usage. If we have reason to use the separate backend-UI structure like the score display, we will separate the two.