Menu Controls Updated
Some unseen glitches in the menu controller input have been fixed. The multi-scaled input is not in use here but the user can mash or hold to cycle through the menu and it feels nice. I suspect that menus in this project will not be too large in scope to warrant a more intricate scrolling system.