Game can now be finished in a way
This commit adds behavior that ends the game and stops the timer once the player crosses the finish line. There are some aspects of the added code that could be rethought or refactored (specifically the way we have two separate lookups for Zones/ZoneTypes) but this makes the player stop cleanly once the finish line is crossed!