Hefty Updates to animation and character orientation!
The code has been cracked on getting smooth animation over curves, the only issue is learning how quaternions actually work in this instance. This is a good opportunity for refactoring any code that uses orientation heavily. In addition, the supremely rough animation system is in, so turning now looks like turning! Animations need big tuning to feel snappy and good.