Major improvements to character alignment!
The lerped alignment is now working for angles which the player is aligned with the surface. This is an exciting improvement! The rotation itself needs to be updated to include surfaces that aren't aligned in a "loop" around the player. The adjustments to the player's position needs toe be re added once the alignment is squared away. From there, issues with player state can be addressed.