Slight improvements to positioning and air/ground bug fixes
The position of the character is now relatively preserved via an interpolation when the surface changes. The result feels a lot better. Also included in this commit are fixes to the player's orientation when catching air and landing -- though there are still vital improvements to be made when jumping from uphill to uphill.