Aerial Movement Refactor pt1
As fixing ticket 9 continues to explode out into a variety of fixes, I think this can safely be considered a hard refactor of the physics system as a whole. This contains the first half of aerial movement refactoring. Transitions need to be fixed, but now the aerial movement is split into the horizontal and vertical components as we would expect in the real world.