Updated landing behavior to be in line with new jumping
The JumpingState class has now been made obsolete in favor of a single aerial state, reducing redundant code. The transition to a grounded state now updates the direction to accommodate for changes in direction relative to the ground direction when the player leaves the ground (key example being quater/half pipe type jumps).