Valuable updates to surface influence
Part of getting the surface influence behavior (slowing down on slopes, etc) down was to implement reversing the players direction. As another part of state to manage, this has complicated the movement mechanics. There are several pretty severe bugs in this WIP commit, but they are relatively easy to diagnose.