More updates to surface influence!
A brainstorm led to the realization that decoupling the transform orientation from the direction of travel allows for more free movement influence, but that is for future development. Currently this commit fixes the "speed up on slowdown" glitch and going uphill does slow down the player. Next steps are to tune the feel of the deceleration and implement angle influence.