Surface Influence works organically!
The surface influence behavior now works in a very logical and good feeling way, with a vector based on the angle of the surface being used to influence the player. This value needs to be fine tuned for good feel, but I am happy with the rough results. However, by including surface influence, I have introduced a regression with stopping, which can easily be solved by having a higher tolerance for what velocity would be considered "stopped." Next step is implementing switch stance using this new method.