Jump transition upgrade and some fixes
The stopping/slowing behavior needed some tuning, as the player never actually stops with surface influence. That has been fixed. The major change here is the updated aerial transition, which now smoothly moves the player from ground to air. The "jerky" jumps have been removed and velocity is no longer increased or decreased based on the direction of travel.