Switch Stance Refactor
The switch stance previously relied on code to cause the change from forward to reverse stance. However, the system can be simplified to a single bool that can only be checked on certain frames (if velocity drops below zero, on landing, etc) and then the animator can handle the switch stance. Ideally, as long as the player's position is preserved when these transitions happen, the player can transition to switch with no issues.