In-Game Pause Menu Implementation
This commit contains a partially functional pause menu for in the middle of the game. It works alongside the rest of the pause system so other entities that have strange pause behavior are not necessarily tied to this new menu. It shows and hides, lets the user iterate over the menu items, and resets when the user hides and reshows. Currently missing is the behavior of each of the times, which will come as part of a refactor effort to implement scalable behavior of menu items.