Character Menu Controller Basics!
The workflow of attaching UI elements to a character select menu that's rendered in a 3D space has been outlined. Placeholders are in place (ha) to test backend functionality, but these will be replaced by cleaner elements. The data updates correctly, so changing text updates to meter alpha updates and image changes is easily handled by the engine.