Spinning basics added, more to come!
The basics for spinning and flipping have been added. Charging, rotating the model, and keeping the camera from freaking out as a result have been added. The camera changes showed the need to break out the movement data into the PlayerPositionData class, though it will remain separate from the model orientation data. Flipping suffers from the same gimbal lock issue that turning on walls has. The numbers need a fair amount of tweaking, too! That will be helped by adding in the analog stick's role in updating the spin rate.