Spinning uni-directional stick influence functioning
Spinning charging now works (in only one direction) with the input axis, with the state machine designed to support Coyote Time in the future. The values are still in need of tweaking but the functionality works in one direction. The next problem to solve is bi-directional spin and flip charges (where changing direction increments and decrements the value accordingly). The solution will likely come in the form of an alternative increment function that allows for absolute value deltas. The gimbal lock shown when turning on angles is exhibited in the flip function as well, so that solution is still out there.