Signed Spinning and the flip/turning gimbal lock has been fixed!
The "gimbal lock" issue was really just redundantly applying the object's orientation to the turn function (and flip function), which resulted in a gimbal-lock-like symptom. Turning is still greatly affected by uphill/downhill, which may or may not be intended, but flipping now works! The only issue is flipping and spinning as a combo need to be changed to work together, as mixing the two merely produces an off axis spin.