Snapping to nearest half/full rotation partial functionality
Currently the player will snap to the nearest half or full rotation when finished spinning. However, there needs to be some post spin adjustment as we only stop rotating after we pass the value, meaning that in most cases the resulting spin will be snapping to the right spot and then a little extra beyond that. Spin/Flip mixtures still need to be tuned up.