Scoring Data now sends correctly and on the right time frame
Fixing up the trick system, the scoring system now sends scoring data to a "judge" as a chunk. Details about the aerials (spinning, grab variety, grab time, air time) are sent and stored in a list, so the score judge can smartly react to it. This judge can be implemented in variety of ways, from a simple "score goes up" system to something more intricate.