Surprise Refactor Effort!
In an effort to start the controls edit menu, I realized I needed to break the controls apart from the key value pairs. So the result became two dictionaries, one referring to the button to check and the other storing the value. This spun out into a huge deal that I'm honestly surprised worked on the first run! Still no idea how to fix that little Quality Edit bug...