Hacky Fix: Quality Edit Controller works better!
Previously, confirming the change on the Quality Edit Controller (and ONLY this controller) preserved the value when the menu controller is reactivated. The result is immediately re-entering the edit controller, or doing nothing in terms of user feedback. This fix manually resets the menu controls on confirmation of any integer value, which has no regressions (as the user presumably doesn't want to make a change and then do something else on the SAME frame) currently shown.