Control edit in... beta! Working with some quirks
The controls edit now works, though the pause button edit is not available. The menu will likely have to be laid out differently to accommodate all the controls. A less-hacky fix has been found for the quality edit controller bug (which was even worse when trying to change the confirm button!). The solution is to set a flag to reset all controller values to IDLE for one frame. Since these actions happen on menus, the effect on game feel should be negligible -- potentially not even perceptible. Still lingering is a similar acting bug but for stepping back one menu further when binding a trick action to the same button as the menu back button.