Major Menu Updates!
The controls menu now fits in 21:9 screens, but the other menus need to be updated. Previously, the text was all sized to be readable in the editor, when the goal should be full screens and televisions. Also included is a fix for the "double back" bug that occurs, as well as removing the potential for race conditions between text menu items and editors below them. I don't think it should be too much trouble to graft this onto an in game settings menu.