Menu Control and UI improvements
The layouts of the menus should now work a little better with ultra wide or ultra thin aspect ratios. Now there is a default dependency on arrow keys and return/backspace to ensure that even if the user turbo-destroys their control set up, they can still reset it reasonably well. This hopefully allows for a wide array of accessible control configurations without the potential of supremely screwing up one's ability to use the game.