The rough cut of the new collision is working!
The player now aligns to the surface and updates rotation accordingly. Though without any position adjustments to adhere to the ground, the new collision check reintroduces a bug in which the player drifts above or below the ground while grounded. This adjustment needs to be thought out. To further improve the ground/air collision, a more robust obstacle collision system needs to be introduced. Also, the velocity-based front raycast needs to be improved.