Extensive bug fixes on the way to a new collision system.
Game-breaking bugs related to quaternions have been ironed out -- at the very least they cannot be reproduced in the current debug build. the front-and-back raycast needs to be tuned up, but it no longer yields a cascading rotation failure. In addition, the introduction of normalized quaternions eliminated a seemingly-random bug in which the offets would (very literally) spiral out of their intended places.