Clipping cases fixed
Rotation has yet to be implemented, but there have been absolutely zero instances of fall-through or clipping that aren't caused by improper wall collision -- that is to say motions that won't be allowed when this is finished. Current outstanding issues: rotation still hasn't been figured out, wall collision and crashing need to be implemented, and sliding off surfaces (angle-limited surface adhesion) needs to be implemented.