Nose and tail checks in place, alignment needs work
The nose and tail, when grounded correctly, now register collisions with the ground at their intended locations. Further tuning may be needed for the front end, but the current major snag is in the ROTATION function, which does not work as intended. The undoing of translations and applying of rotations needs to be thought out to get the right alignments, because right now things work okay when the alignment is right!