Ground collision reorientation working!
Inspired by SSX3: apparently no matter the speed or angle, a ground-based collision with an obstacle just leads to slowdown and a reorientation. The reorientation works, "crashing" is now something that no longer happens (but will, of course, when tricks are failed). The function works when it occurs, however it assumes the player's orientation is correct, which still needs implementation.