Quadratics in place, buggy currently
While there are still kinks to work out with how this works, the core idea of finding the intersection of our travel plane and the target plane, then using a quadratic function to find two possible new nose points, then deciding between the two is in place. This does not currently work, but I am satisfied with the progress as the numbers imply that the issue is minor.