Quadratics working, heuristic and obstacle collision need work
The quadratics have been fixed up, verifying that the rotation of the position and direction work correctly. A few minor issues, unrelated to the basics, remain: first the heuristic for deciding "up" or "down" is very wrong, it needs work later. Next, the obstacle detection appears to not work correctly with a rotated orientation. There is a "boost" and a "wobble" once we leave the base orientation. A good goal to shoot for next is to ensure that as long as the tail and nose are coplanar, no rotation occurs. The wonky heuristic might be an issue.