Angle heuristic improved, new goals in place
There are a few goals that need to be reached before this is finalized. First is the axis alignment: when hitting a new slope at an angle, there is a misalignment that needs to be adjusted. Second is overall settling: the initial turns seem to work, but if there's some kind of axis mismatch that should be smoothed out. Third is clipping: we still experience clipping into the ground (which is slightly intended, though not nearly to this degree). Fourth is aerial transition: how to prevent sufficiently steep slopes from grounding the player.