More progress and new ideas implemented
The angle calculation doesn't appear to go all the way when we want it to, but having the source of the issue is still progress toward fixing it. I suspect this, mixed with the normal not getting rotated at all, leads to the weird desyncs. There are problems at hand but the path forward is there.