Drastic improvement thanks to barycentric coordinates!
Overall, movement along curved surfaces is a lot better thanks to the use of a new surface adjustment function and the use of barycentric coordinates to simplify the new normal calculation. The end result works nicely in straight lines, but the reorientation calculation breaks turning. That's isolated and simple, so overall this is a major improvement.