Normal circumstances are looking great!
There is NO wall collision detection. If the collision detection system detects a wall, there is no behavior defined currently. This is important because currently ground collisions are by and large in place. I cannot get a fall-through case provided there wasn't some unrelated error before. In addition, playing with normal smoothing has given us a path to much better environments and interactions. The current issue is jittering based on the function used to keep the player above the ground, but that can be ironed out over time.