Psyche! Previous commit was a lie
Let this be a lesson in always BUILDING before confirming something. Due to the position being on the player and not the board, the rotations in the "real" game went haywire and the game was unplayable. The solution was to move the position to be on the board rather than on the body, as the body is really just along for the ride (at least when determining if we're grounded or aerial). Several rough-edits are included, but the positioning of raycasts will be cleaned up as this gets refined. Outstanding is the issue of when we get a surface normal but no other contact points (like when we hit steep walls JUST right) and still no zap-to-surface adjustment.