Rough edit of the countdown added!
The pre-event countdown has been added. Currently the counter only does two seconds instead of three. As well, no UI has been implemented currently, only a start when the user presses a button and the "pre-start" state for the player (not the camera or other game objects). These will be added, but the core math is good!