Pre-countdown preview loosely implemented!
The UI still needs flavor and spice as it is all just text, but now before the countdown there is a system for tracking shots of the level that is completely modular. The current issues are the camera's need to translate from wherever it is TO the start position as well as the UI being the in-game UI rather than some preview prompt. However, that stuff feels like it can easily be wrapped in the addition of the countdown timer.