Quick UI organization boost
This is a path-of-least-resistance update, since the preview occurs once at the start of the screen and never again. Using more complex data structures for something that gets deleted almost immediately seems like overengineering, so caching the active menu and enabling the menu we want and disabling the menu we don't works just fine.