Memory Allocation improvements
The garbage collector ultimately reports fewer allocations, evidently zero. Though the inconsistencies between player and editor make me wonder. That's for later, as we now have the functionality in place to update a countdown image! This could further be improved by adding positions or layering the images together, but the updates work and don't seem to break anything. The images are currently placeholders, obviously. They will soon be replaced with more adequate placeholders.