Basic Surface Adjustment working!
A key problem with the surface adjustment is it is not tied to state, which breaks design conventions we currently have. The iState interface needs to be rethought for the possibility of SOME but NOT ALL states wanting an Act() for FixedUpdate() in addition to Act() in Update(). Still not entirely perfect, but I suspect that this is going to work great with a well-made environment.