In progress: jumping actually working in player
The current issue is the jump doesn't work! Due to what I assume is a script execution order issue, when we try to jump, it registers the first frame but lands immediately. Changing the distance of "bumping" the player off the surface does not fix it, so the way Update and FixedUpdate() relate need to be changed.