Improvement to physics systems - framerate independence
Previously, Update was used for most movement while FixedUpdate was used for the collision. Due to the "mismatch" here, it felt jittery and there wasn't a clean way to guarantee jumps would register. The solution is to only use FixedUpdate for game logic (with the added benefit of better determinism) and interpolate the "display" position between FixedUpdate calls. Input needs to be refactored as some inputs get lost... but when it works it works great!