Input on track with framerate independence!
Using a last-frame-current-frame function, the safe-in-FixedUpdate Input.GetKey() is now used to get inputs that appear to not drop any frames. I believe that this results in control that is purely framerate independent. Notable outstanding issues - the camera needs to be reworked as the transform rotation can't be used, neither can the "backend" values, so FixedUpdate will be used for the camera. Also, if the player's velocity hits a negative value, jumping and landing can cause an orientation issue that leads to fall-through. These are relatively easy to debug and fix, I believe! Famous last words :)