Center of gravity update (and pause catcher)!
Previously, the movement of the position to be a function of the board instead of the player's body (as this makes the most logical sense on the ground). This caused flips in the air to look funky, as it appears as though the center of gravity was the board and not the player. This commit updates the rotation setup to allow for a separate center of gravity. Also, pauses could get "dropped" due to the Update/FixedUpdate mismatch, but this has been fixed.