Acceleration added to turning for weight
Previously, turning was a function of whatever the current input axis value was. The result was snappy, weightless turning. Now, an acceleration value has been incorporated into handling which gives the handling a little weight. The numbers need to be balanced but it feels better having weight than having none!