Camera behavior update beginning!
The camera behavior has now been updated to use the new interpolated movement methodology. However, almost everything about the camera needs to be rethought as the way the player moves has changed drastically. It might be easier to gut the thing entirely and try again, but we'll think it over!