(UNTESTED CHANGES) More basics
A conceptually sound (I think!) version of the vertical motion has been added. This should make it so if the camera reaches a certain angle, it only translates, otherwise it merely rotates to stay focused on the player. However, the sign of the rotation needs to be reviewed and the effects of other translations needs to be thought out.