Camera behavior improved!
The camera goes to lightspeed and disappears. That makes sense, as camera focus has not been implemented. However, transitioning to following the player works and the translation works as long as the surface is flat. It is jittery, which can be fixed by managing Update() and FixedUpdate(), The basics are continuing to come together, piece by piece.