Odds and ends!
The angle slowdown/speedup was NOT time dependent and thus needed a deltaTime. In addition, the PSX style rendering has been added as well as the low resolution render texture. The environment will be spruced up over time and animations will come. The primary goal currently is a playable, shareable demo.