Aerial Camera bug fixed!
The camera previously used the player-message-provided rotation to calculate the vertical and horizontal translation, which led it to lag behind when in the air (where the player's rotation is constant). This was changed to just use the current camera rotation. In addition, the velocity cap has been ignored, allowing the camera to "elastically" stay close to the player.