Level-informed physics updates
The previous top speed adjustment code was bizarre in retrospect, using gravity when terminal velocity should be used to inform top speed, as well as having top speed be a hard cap when it should be a soft cap. The soft cap calculation still needs to be thought out, but it feels much better now. Turn tuning as well!