Collision Refinements
The return of the singular raycast to define direction! It's simple but it does a great job, as long as there's a sufficiently large "bump" to keep the player above the ground. What is currently in need of fixing is some refinements to raycast distance and seeing why there's the occasional "bounce"