Simultaneous Input Achieved!
Keyboard and Controller Input can now bee freely alternated. What is currently missing is input masks to only allow rebinding of keyboard keys to the keyboard and controller keys to the controller. That's busywork and will come soon.

This could pose issues down the road when trying to implement multiplayer, but that is out of the scope of this project. That can easily be remedied by adding a separate input controller and disabling one of the two input sources.