Updated Input Options
The list of valid keyboard keys has been expanded to include common keys on a keyboard. This ideally would include every possible key in Unity's KeyCode manager but this should likely be sufficient for the overwhelming majority of uses. In addition, now only keyboard keys can be bound to the keyboard and the same for controller input. Default setting has also been fixed.

The next major steps are reliably getting the last used input device and updating UI accordingly, and down the road allowing for axes to be bound as keys.