Messy hot swapping update
This commit mostly adds the UI automatically updating when the input device changes on the fly. However the changes have introduced a regression with the global data controller that unbinds and "falsely" binds the keys on next startup. That will be fixed in a coming commit.

Ultimately I am concerned with the implementation here, it works but uses some "back and forth" that seems prone to errors down the road. Luckily there are few places where this could become an issue, so hopefully that bodes well for the future.