Turn Slowing implemented
Now the player slightly slows when turning, to make it a deliberate choice to make and encourage ideal lines. The numbers may be overtuned right now, so that will change in time. Also included is some code cleanup moving carts to static space, as well making turn acceleration capped by how much the stick is held, reducing an unwanted behavior in which even a light tilt of the stick would eventually cause a full on turn.