Aerial Bug fixes
In a step in the right direction, now when the player returns to the ground at an awkward angle, the player will not clip through the ground but instead will bounce in some cases, land cleanly in others. This can be debugged and is ultimately preferable. Also included is the same speed decay on wall collisions when in the air. 

What currently needs fixing is the jump behavior still seems awkward, there is a "hitch" that pushes the player forward and mathematically it is confusing. That needs to be addressed alongside finding a way to guarantee liftoff when jumping and not immediately grounding on the next frame.