New Time Scale and New Collision Implementation
One of the major improvements, despite all the flaws, is that the player has not ONCE clipped through the ground using OverlapBoxNonAlloc(). However, there is some "hitching" based on how we decide we're entering the air. The hitches are probably due to the wonky aerial transition, but the heuristic for changing the grounded state to "not grounded" needs to be updated.