Significantly improved ground collision
The ground collision works pretty perfectly on concave curves, but convex curves and steep slopes still introduce some hitching. In addition, some change here appears to have broken the player's model rotation on land, which will reset to the direction of travel when landing even when it shouldn't. Massive improvements, excited to be reaching the nitpicky stage!