Camera Update started!
The old camera can get hitchy and is far too closely coupled to the player's state. It also had some really strange edge cases that caused it to completely freak out. As a result, we are refactoring it (most importantly ignoring player rotation) to work as smoothly as the new collision does!