Boost state started!
A boost state has been added, that allows the player to move faster at the expense of slower turns and being unable to jump. Some reference material is needed to make sure this is applied correctly. The hard speed cap has been changed to a soft cap, and the next step in this process is the "boost jump cancel" and adding in an acceleration bump to the boost. A boost meter may be a future mechanic