We've got boost power!
Boosting has now been fully implemented, at least as a "crouch" mechanic a la 1080 Avalanche. A metered boost can easily be tacked on depending on what mechanics we want to focus on. The boost is scaled so steeper slopes allow more boost and the boost can't be used for climbing up hill. In addition, it extends the top speed cap and gives an acceleration bump. Also it cancels jump charging!