The second twist in the braid
The three-path over under knot thing has a second part to it ready to go. One more "tie" that looks a bit different to close it out, then we will have an ending turn. After that, the only thing left is just adding in side space to fill it out, then trees and fencing for design reasons! We're almost there!