Nicely updated ground rotation adjustment
Previously, the "ground check" would break if the change in angle between the player's launch point and land point was sufficiently large. Doing a direct downward cast seems to work, though it's somewhat buggy (though the bugs are strange-looking rather than the previous gamebreakers). A potential issue is wholly vertical slopes, so an alternative check may be helpful. Perhaps a context sensitive function (player orientation when grounded, directly down in the air)?